+++ i7dw/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.7.3"
+__version__ = "0.7.4"
 
 import logging
