+++ i7dw/interpro/elastic/relationship.py
@@ -4,72 +4,15 @@ import hashlib
 import time
 from multiprocessing import Process, Queue
 from tempfile import mkdtemp
-from typing import Dict, List
+from typing import List
 
 from . import index, set_ready
-from .. import mysql, repr_frag
+from .. import condense, mysql
 from ... import logger, pdbe
 from ...io import JsonFileOrganizer, Store
 
 
 NODB_INDEX = "others"
-
-
-def condense(to_condense: Dict[str, List]) -> Dict[str, List]:
-    condensed = {}
-    for entry_ac, locations in to_condense.items():
-        start = end = None
-        _locations = []
-
-        # Sort location by the position of the leftmost fragment
-        for frags in sorted(locations, key=lambda l: repr_frag(l[0])):
-            """
-            We do not consider fragmented matches:
-                - `s` is the leftmost start position
-                - `e` is the rightmost end position
-                (assuming `frags` is sorted by (start, end) keys)
-            """
-            s = frags[0]["start"]  # leftmost start position
-            e = frags[-1]["end"]  # rightmost end position
-
-            if start is None:
-                start = s
-                end = e
-            elif s > end:
-                """
-                      end
-                   [----] [----]
-                          s
-                -> new location
-                """
-                _locations.append((start, end))
-                start = s
-                end = e
-            elif e > end:
-                """
-                        end
-                   [----]
-                     [------]
-                            e
-                -> extend
-                """
-                end = e
-
-        _locations.append((start, end))
-
-        condensed[entry_ac] = []
-        for start, end in _locations:
-            condensed[entry_ac].append({
-                "fragments": [{
-                    "start": start,
-                    "end": end,
-                    "dc-status": "CONTINUOUS"
-                }],
-                "seq_feature": None,
-                "model_acc": None
-            })
-
-    return condensed
 
 
 class DocumentProducer(Process):
